앱 완성

+++ AndroidApp/app/src/main/java/kr/ac/kumoh/ce/s20190348/AndroidAppProject/Sensor.kt
@@ -1,7 +1,7 @@
 package kr.ac.kumoh.ce.s20190348.AndroidAppProject
 
 data class Sensor(
-    val temperature : Int,
+    val temperature : Float,
     val water : Int,
     val gas : Int,
     val nfc : String,
